feat: refactor dialog component
