+++ App.tsx
@@ -4,7 +4,7 @@ import { SafeAreaProvider } from 'react-native-safe-area-context'
 import Toast from 'react-native-toast-message'
 
 import Navigation from '@navigation/index'
-import { DialogProvider } from '@context/dialog'
+import { DialogProvider } from '@context/dialog.context'
 import { store } from '@store/index'
 import toastConfig from '@config/toast.config'
 import AppInitializer from '@root/app-initializer'
